fix(cli): improve scanner error handling and refresh dist
